+++ samples/WebAPI_Dapper/build.sh
@@ -1,0 +1,15 @@
+pushd GreetingsWeb || exit
+rm -rf out
+dotnet restore
+dotnet build
+dotnet publish -c Release -o out
+docker build .
+popd || exit
+pushd SalutationAnalytics || exit
+rm -rf out
+dotnet restore
+dotnet build
+dotnet publish -c Release -o out
+docker build .
+popd || exit
+