+++ release_notes.md
@@ -37,6 +37,7 @@ This section lists features in master, available by [AppVeyor](https://ci.appvey
  - Provided a short form of the BrighterMessaging constructor, that queries object provided for async versions of interfaces
  - Changed IsAsync to RunAsync on a Subscription for clarity
  - Supports an async pipeline: callbacks should happen on the same thread as the handler (and the pump), avoiding thread pool threads
+ - Fixed issue in SQlite with SQL to mark a message as dispatched
  
 ## Release 8.1.1399 ##
  - Update nuget libs